Refactor: Extract out GRPC client
Moving GRPC client logic into a shared instance to prepare for
deduplication.

Removing the Predicate pushdown from the LookupSource. This is no longer
desired as the (de)serialization shapes would be different, reducing the
ability to deduplicate requests.
